refactor: bottom sheet ref
