+++ components/BottomSheet/index.ts
@@ -1,2 +1,1 @@
 export { default as BottomSheet } from './BottomSheet'
-export { default as bottomSheetRef } from './ref'
